quick-start: update sample to New Installation Methods
