+++ quick-start/node-server/editor/package.json
@@ -5,30 +5,11 @@
   "version": "0.0.1",
   "license": "SEE LICENSE IN LICENSE.md",
   "private": true,
+  "dependencies": {
+    "ckeditor5": "*",
+    "ckeditor5-premium-features": "*"
+  },
   "devDependencies": {
-    "@ckeditor/ckeditor5-autoformat": "*",
-    "@ckeditor/ckeditor5-basic-styles": "*",
-    "@ckeditor/ckeditor5-block-quote": "*",
-    "@ckeditor/ckeditor5-cloud-services": "*",
-    "@ckeditor/ckeditor5-comments": "*",
-    "@ckeditor/ckeditor5-dev-utils": "43.0.0",
-    "@ckeditor/ckeditor5-dev-webpack-plugin": "*",
-    "@ckeditor/ckeditor5-easy-image": "*",
-    "@ckeditor/ckeditor5-editor-classic": "*",
-    "@ckeditor/ckeditor5-essentials": "*",
-    "@ckeditor/ckeditor5-heading": "*",
-    "@ckeditor/ckeditor5-image": "*",
-    "@ckeditor/ckeditor5-indent": "*",
-    "@ckeditor/ckeditor5-link": "*",
-    "@ckeditor/ckeditor5-list": "*",
-    "@ckeditor/ckeditor5-media-embed": "*",
-    "@ckeditor/ckeditor5-paragraph": "*",
-    "@ckeditor/ckeditor5-paste-from-office": "*",
-    "@ckeditor/ckeditor5-real-time-collaboration": "*",
-    "@ckeditor/ckeditor5-table": "*",
-    "@ckeditor/ckeditor5-track-changes": "*",
-    "@ckeditor/ckeditor5-theme-lark": "*",
-    "@ckeditor/ckeditor5-typing": "*",
     "css-loader": "^5.2.7",
     "postcss": "^8.4.12",
     "postcss-loader": "^4.3.0",
