+++ quick-start/node-server/editor/sample/main.js
@@ -1,0 +1,86 @@
+/**
+ * @license Copyright (c) 2014-2025, CKSource Holding sp. z o.o. All rights reserved.
+ * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
+ */
+import 'ckeditor5/ckeditor5.css';
+import 'ckeditor5-premium-features/ckeditor5-premium-features.css';
+
+import {
+	ClassicEditor,
+	Essentials,
+	Autoformat,
+	BlockQuote,
+	Bold,
+	CloudServices,
+	List,
+	EasyImage,
+	Heading,
+	Image,
+	ImageCaption,
+	ImageStyle,
+	ImageToolbar,
+	ImageUpload,
+	Indent,
+	Italic,
+	MediaEmbed,
+	Link,
+	Paragraph,
+	PasteFromOffice,
+	Table,
+	TableToolbar,
+	TextTransformation,
+	Underline
+} from 'ckeditor5';
+
+import {
+	Comments,
+	PresenceList,
+	RealTimeCollaborativeComments,
+	RealTimeCollaborativeEditing,
+	RealTimeCollaborativeTrackChanges,
+	TrackChanges
+} from 'ckeditor5-premium-features';
+
+class Editor extends ClassicEditor {}
+
+// Plugins to include in the build.
+Editor.builtinPlugins = [
+	Autoformat,
+	BlockQuote,
+	Bold,
+	EasyImage,
+	Essentials,
+	Heading,
+	Image,
+	ImageCaption,
+	ImageStyle,
+	ImageToolbar,
+	ImageUpload,
+	Indent,
+	Italic,
+	Link,
+	List,
+	MediaEmbed,
+	Paragraph,
+	PasteFromOffice,
+	PresenceList,
+	RealTimeCollaborativeComments,
+	RealTimeCollaborativeEditing,
+	RealTimeCollaborativeTrackChanges,
+	Table,
+	TableToolbar,
+	TextTransformation,
+	Comments,
+	TrackChanges,
+	CloudServices
+];
+
+Editor.defaultConfig = {
+	comments: {
+		editorConfig: {
+			extraPlugins: [ Bold, Italic, Underline, List, Autoformat ]
+		}
+	}
+};
+
+export default Editor;
