--- quick-start/node-server/editor/src/ckeditor.js
@@ -1,78 +1,0 @@
-/**
- * @license Copyright (c) 2014-2022, CKSource Holding sp. z o.o. All rights reserved.
- * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
- */
-import ClassicEditor from '@ckeditor/ckeditor5-editor-classic/src/classiceditor.js';
-import Autoformat from '@ckeditor/ckeditor5-autoformat/src/autoformat.js';
-import BlockQuote from '@ckeditor/ckeditor5-block-quote/src/blockquote.js';
-import Bold from '@ckeditor/ckeditor5-basic-styles/src/bold.js';
-import EasyImage from '@ckeditor/ckeditor5-easy-image/src/easyimage.js';
-import Essentials from '@ckeditor/ckeditor5-essentials/src/essentials.js';
-import Heading from '@ckeditor/ckeditor5-heading/src/heading.js';
-import Image from '@ckeditor/ckeditor5-image/src/image.js';
-import ImageCaption from '@ckeditor/ckeditor5-image/src/imagecaption.js';
-import ImageStyle from '@ckeditor/ckeditor5-image/src/imagestyle.js';
-import ImageToolbar from '@ckeditor/ckeditor5-image/src/imagetoolbar.js';
-import ImageUpload from '@ckeditor/ckeditor5-image/src/imageupload.js';
-import Indent from '@ckeditor/ckeditor5-indent/src/indent.js';
-import Italic from '@ckeditor/ckeditor5-basic-styles/src/italic.js';
-import Link from '@ckeditor/ckeditor5-link/src/link.js';
-import List from '@ckeditor/ckeditor5-list/src/list.js';
-import MediaEmbed from '@ckeditor/ckeditor5-media-embed/src/mediaembed.js';
-import Paragraph from '@ckeditor/ckeditor5-paragraph/src/paragraph.js';
-import PasteFromOffice from '@ckeditor/ckeditor5-paste-from-office/src/pastefromoffice';
-import PresenceList from '@ckeditor/ckeditor5-real-time-collaboration/src/presencelist.js';
-import RealTimeCollaborativeComments from '@ckeditor/ckeditor5-real-time-collaboration/src/realtimecollaborativecomments.js';
-import RealTimeCollaborativeEditing from '@ckeditor/ckeditor5-real-time-collaboration/src/realtimecollaborativeediting.js';
-import RealTimeCollaborativeTrackChanges from '@ckeditor/ckeditor5-real-time-collaboration/src/realtimecollaborativetrackchanges.js';
-import Table from '@ckeditor/ckeditor5-table/src/table.js';
-import TableToolbar from '@ckeditor/ckeditor5-table/src/tabletoolbar.js';
-import TextTransformation from '@ckeditor/ckeditor5-typing/src/texttransformation.js';
-import Underline from '@ckeditor/ckeditor5-basic-styles/src/underline';
-import Comments from '@ckeditor/ckeditor5-comments/src/comments';
-import TrackChanges from '@ckeditor/ckeditor5-track-changes/src/trackchanges';
-import CloudServices from '@ckeditor/ckeditor5-cloud-services/src/cloudservices';
-
-class Editor extends ClassicEditor {}
-
-// Plugins to include in the build.
-Editor.builtinPlugins = [
-	Autoformat,
-	BlockQuote,
-	Bold,
-	EasyImage,
-	Essentials,
-	Heading,
-	Image,
-	ImageCaption,
-	ImageStyle,
-	ImageToolbar,
-	ImageUpload,
-	Indent,
-	Italic,
-	Link,
-	List,
-	MediaEmbed,
-	Paragraph,
-	PasteFromOffice,
-	PresenceList,
-	RealTimeCollaborativeComments,
-	RealTimeCollaborativeEditing,
-	RealTimeCollaborativeTrackChanges,
-	Table,
-	TableToolbar,
-	TextTransformation,
-	Comments,
-	TrackChanges,
-	CloudServices
-];
-
-Editor.defaultConfig = {
-	comments: {
-		editorConfig: {
-			extraPlugins: [ Bold, Italic, Underline, List, Autoformat ]
-		}
-	}
-};
-
-export default Editor;
