+++ quick-start/node-server/editor/webpack.config.js
@@ -1,5 +1,5 @@
 /**
- * @license Copyright (c) 2014-2022, CKSource Holding sp. z o.o. All rights reserved.
+ * @license Copyright (c) 2014-2025, CKSource Holding sp. z o.o. All rights reserved.
  * For licensing, see LICENSE.md or https://ckeditor.com/legal/ckeditor-oss-license
  */
 
@@ -8,31 +8,27 @@
 /* eslint-env node */
 
 const path = require( 'path' );
-const webpack = require( 'webpack' );
-const { bundler, styles } = require( '@ckeditor/ckeditor5-dev-utils' );
-const CKEditorWebpackPlugin = require( '@ckeditor/ckeditor5-dev-webpack-plugin' );
-const TerserWebpackPlugin = require( 'terser-webpack-plugin' );
+const TerserPlugin = require( 'terser-webpack-plugin' );
 
 module.exports = {
 	devtool: 'source-map',
 	performance: { hints: false },
+	mode: 'production',
 
-	entry: path.resolve( __dirname, 'src', 'ckeditor.js' ),
+	entry: path.resolve( __dirname, 'sample', 'main.js' ),
 
 	output: {
 		// The name under which the editor will be exported.
-		library: 'ClassicEditor',
-
-		path: path.resolve( __dirname, 'build' ),
-		filename: 'ckeditor.js',
+		filename: 'editor.bundle.js',
+		library: 'CKEditorCS',
+		path: path.resolve( __dirname, 'dist' ),
 		libraryTarget: 'umd',
 		libraryExport: 'default'
 	},
 
 	optimization: {
 		minimizer: [
-			new TerserWebpackPlugin( {
-				sourceMap: true,
+			new TerserPlugin( {
 				terserOptions: {
 					output: {
 						// Preserve CKEditor 5 license comments.
@@ -44,25 +40,8 @@ module.exports = {
 		]
 	},
 
-	plugins: [
-		new CKEditorWebpackPlugin( {
-			// UI language. Language codes follow the https://en.wikipedia.org/wiki/ISO_639-1 format.
-			// When changing the built-in language, remember to also change it in the editor's configuration (src/ckeditor.js).
-			language: 'en',
-			additionalLanguages: 'all'
-		} ),
-		new webpack.BannerPlugin( {
-			banner: bundler.getLicenseBanner(),
-			raw: true
-		} )
-	],
-
 	module: {
 		rules: [
-			{
-				test: /\.svg$/,
-				use: [ 'raw-loader' ]
-			},
 			{
 				test: /\.css$/,
 				use: [
@@ -77,17 +56,6 @@ module.exports = {
 					},
 					{
 						loader: 'css-loader'
-					},
-					{
-						loader: 'postcss-loader',
-						options: {
-							postcssOptions: styles.getPostCssConfig( {
-								themeImporter: {
-									themePath: require.resolve( '@ckeditor/ckeditor5-theme-lark' )
-								},
-								minify: true
-							} )
-						}
 					}
 				]
 			}
